Move FileWatchedPortableExecutableReferenceFactory to the workspaces layer
It's not clear to me why I made this a MEF service when just a simple
constructable type works just fine.
